+++ ovenproof_scoreboard_plugin/scripts/mods/ovenproof_scoreboard_plugin/ovenproof_scoreboard_plugin_localization.lua
@@ -1,7 +1,7 @@
 local mod = get_mod("ovenproof_scoreboard_plugin")
 local UIRenderer = mod:original_require("scripts/managers/ui/ui_renderer")
 local ui_renderer_instance = Managers.ui:ui_constant_elements():ui_renderer()
-local languages = {"en","ru","zh-cn", "zh-tw"}
+local languages = {"en","ru","zh-cn","zh-tw"}
 
 mod.get_text_size = function(self, input_text)