+++ docs/CHANGELOG.md
@@ -11,3 +11,4 @@ v1.2.0-beta-branch
     - Mutator disablers (Grandfather's Gifts) from specialists to disablers. Thanks Tunnfisk!
     - Fix for bleed and warpfire damage counting as melee (removing buff from melee type. thanks syllogism!)
     - Moved shockmaul_stun to dog damage type, since shock maul electricity damage is less important than dog shocks (thanks for the suggestion syllogism!). Planning on a "cleaner" solution to this later
+- Coding style for the localizations